+++ src/nock.utils.test.ts
@@ -2,11 +2,11 @@ import { ApiKeyPlacement, HttpScheme, SecurityType } from '@superfaceai/ast';
 
 import { RecordingDefinition } from '.';
 import {
-  HIDDEN_CREDENTIALS_PLACEHOLDER,
-  HIDDEN_PARAMETERS_PLACEHOLDER,
   replaceCredentialInDefinition,
   replaceParameterInDefinition,
 } from './nock.utils';
+
+const TMP_PLACEHOLDER = 'placeholder';
 
 describe('nock utils', () => {
   describe('removeCredentials', () => {
@@ -32,6 +32,7 @@ describe('nock utils', () => {
           },
           baseUrl: 'https://localhost',
           credential: 'secret',
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
@@ -40,7 +41,7 @@ describe('nock utils', () => {
           method: 'GET',
           status: 200,
           reqheaders: {
-            ['api_key']: HIDDEN_CREDENTIALS_PLACEHOLDER,
+            ['api_key']: TMP_PLACEHOLDER,
           },
         });
       });
@@ -96,6 +97,7 @@ describe('nock utils', () => {
           },
           baseUrl: 'https://localhost',
           credential: 'secret',
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
@@ -104,7 +106,7 @@ describe('nock utils', () => {
           method: 'GET',
           status: 200,
           body: {
-            my_api_key: HIDDEN_CREDENTIALS_PLACEHOLDER,
+            my_api_key: TMP_PLACEHOLDER,
           },
         });
       });
@@ -127,11 +129,12 @@ describe('nock utils', () => {
           },
           baseUrl: 'https://localhost',
           credential: 'secret',
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
           scope: 'https://localhost',
-          path: `/get/${HIDDEN_CREDENTIALS_PLACEHOLDER}?text=123`,
+          path: `/get/${TMP_PLACEHOLDER}?text=123`,
           method: 'GET',
           status: 200,
         });
@@ -155,11 +158,12 @@ describe('nock utils', () => {
           },
           baseUrl: 'https://gitlab.com/api', //Path ends with /api
           credential: 'secret',
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
           scope: 'https://localhost',
-          path: `/api/v4/get/${HIDDEN_CREDENTIALS_PLACEHOLDER}?text=123`,
+          path: `/api/v4/get/${TMP_PLACEHOLDER}?text=123`,
           method: 'GET',
           status: 200,
         });
@@ -183,11 +187,12 @@ describe('nock utils', () => {
           },
           baseUrl: 'https://localhost',
           credential: 'secret',
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
           scope: 'https://localhost',
-          path: `/get?api_key=${HIDDEN_CREDENTIALS_PLACEHOLDER}&text=123`,
+          path: `/get?api_key=${TMP_PLACEHOLDER}&text=123`,
           method: 'GET',
           status: 200,
         });
@@ -257,6 +262,7 @@ describe('nock utils', () => {
           },
           baseUrl: 'https://localhost',
           credential: 'secret',
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
@@ -265,7 +271,7 @@ describe('nock utils', () => {
           method: 'GET',
           status: 200,
           reqheaders: {
-            ['Authorization']: `Basic ${HIDDEN_CREDENTIALS_PLACEHOLDER}`,
+            ['Authorization']: `Basic ${TMP_PLACEHOLDER}`,
           },
         });
       });
@@ -368,6 +374,7 @@ describe('nock utils', () => {
           },
           baseUrl: 'https://localhost',
           credential: 'secret',
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
@@ -376,7 +383,7 @@ describe('nock utils', () => {
           method: 'GET',
           status: 200,
           reqheaders: {
-            ['Authorization']: `Bearer ${HIDDEN_CREDENTIALS_PLACEHOLDER}`,
+            ['Authorization']: `Bearer ${TMP_PLACEHOLDER}`,
           },
         });
       });
@@ -479,6 +486,7 @@ describe('nock utils', () => {
           definition,
           baseUrl,
           credential: parameterValue,
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
@@ -508,6 +516,7 @@ describe('nock utils', () => {
           definition,
           baseUrl,
           credential: parameterValue,
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
@@ -516,7 +525,7 @@ describe('nock utils', () => {
           method: 'GET',
           status: 200,
           reqheaders: {
-            ['api_key']: HIDDEN_PARAMETERS_PLACEHOLDER,
+            ['api_key']: TMP_PLACEHOLDER,
           },
         });
       });
@@ -537,6 +546,7 @@ describe('nock utils', () => {
           definition,
           baseUrl,
           credential: parameterValue,
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
@@ -545,7 +555,7 @@ describe('nock utils', () => {
           method: 'GET',
           status: 200,
           body: {
-            my_api_key: HIDDEN_PARAMETERS_PLACEHOLDER,
+            my_api_key: TMP_PLACEHOLDER,
           },
         });
       });
@@ -563,11 +573,12 @@ describe('nock utils', () => {
           definition,
           baseUrl,
           credential: parameterValue,
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
           scope: baseUrl,
-          path: `/get/${HIDDEN_PARAMETERS_PLACEHOLDER}?text=123`,
+          path: `/get/${TMP_PLACEHOLDER}?text=123`,
           method: 'GET',
           status: 200,
         });
@@ -586,11 +597,12 @@ describe('nock utils', () => {
           definition,
           baseUrl,
           credential: parameterValue,
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
           scope: baseUrl,
-          path: `/get?api_key=${HIDDEN_PARAMETERS_PLACEHOLDER}&text=123`,
+          path: `/get?api_key=${TMP_PLACEHOLDER}&text=123`,
           method: 'GET',
           status: 200,
         });
@@ -610,10 +622,11 @@ describe('nock utils', () => {
           definition,
           baseUrl,
           credential: parameterValue,
+          placeholder: TMP_PLACEHOLDER,
         });
 
         expect(definition).toEqual({
-          scope: `https://api.${HIDDEN_PARAMETERS_PLACEHOLDER}.com`,
+          scope: `https://api.${TMP_PLACEHOLDER}.com`,
           path: `/get?text=123`,
           method: 'GET',
           status: 200,