+++ src/nock.utils.ts
@@ -15,8 +15,7 @@ import { UnexpectedError } from './common/errors';
 interface ReplaceOptions {
   definition: RecordingDefinition;
   credential: string;
-  isParameter: boolean;
-  placeholder?: string;
+  placeholder: string;
 }
 
 const debug = createDebug('superface:testing:recordings');
@@ -29,37 +28,23 @@ CONSIDER DISABLING SENSITIVE INFORMATION LOGGING BY APPENDING THE DEBUG ENVIRONM
 `
 );
 
-export const HIDDEN_CREDENTIALS_PLACEHOLDER =
-  'credentials-removed-to-keep-them-secure';
-export const HIDDEN_PARAMETERS_PLACEHOLDER =
-  'parameters-removed-to-keep-them-secure';
 const AUTH_HEADER_NAME = 'Authorization';
-
-const defaultPlaceholder = (isParameter: boolean) =>
-  isParameter ? HIDDEN_PARAMETERS_PLACEHOLDER : HIDDEN_CREDENTIALS_PLACEHOLDER;
 
 function replaceCredential({
   payload,
   credential,
-  isParameter,
   placeholder,
 }: {
   payload: string;
   credential: string;
-  isParameter: boolean;
-  placeholder?: string;
+  placeholder: string;
 }) {
   if (credential !== '') {
     debugSensitive(
-      `Replacing credential: '${credential}' for placeholder: '${
-        placeholder ?? defaultPlaceholder(isParameter)
-      }'`
+      `Replacing credential: '${credential}' for placeholder: '${placeholder}'`
     );
 
-    return payload.replace(
-      new RegExp(credential, 'g'),
-      placeholder ?? defaultPlaceholder(isParameter)
-    );
+    return payload.replace(new RegExp(credential, 'g'), placeholder);
   }
 
   return payload;
@@ -68,7 +53,6 @@ function replaceCredential({
 function replaceCredentialInHeaders({
   definition,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions): void {
   if (definition.reqheaders) {
@@ -84,7 +68,6 @@ function replaceCredentialInHeaders({
       definition.reqheaders[headerName] = replaceCredential({
         payload: headerValue.toString(),
         credential,
-        isParameter,
         placeholder,
       });
     }
@@ -94,7 +77,6 @@ function replaceCredentialInHeaders({
 function replaceCredentialInRawHeaders({
   definition,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions): void {
   if (definition.rawHeaders) {
@@ -106,7 +88,6 @@ function replaceCredentialInRawHeaders({
       return replaceCredential({
         payload: header,
         credential,
-        isParameter,
         placeholder,
       });
     });
@@ -116,30 +97,29 @@ function replaceCredentialInRawHeaders({
 function replaceCredentialInBody({
   definition,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions): void {
-  if (definition.body !== undefined) {
+  if (definition.body !== undefined && definition.body !== '') {
     let body = JSON.stringify(definition.body);
 
-    debug('Replacing credentials in request body');
-    debugSensitive('Request body:', body);
+    if (body.includes(credential)) {
+      debug('Replacing credentials in request body');
+      debugSensitive('Request body:', body);
 
-    body = replaceCredential({
-      payload: body,
-      credential,
-      isParameter,
-      placeholder,
-    });
+      body = replaceCredential({
+        payload: body,
+        credential,
+        placeholder,
+      });
 
-    definition.body = JSON.parse(body) as RequestBodyMatcher;
+      definition.body = JSON.parse(body) as RequestBodyMatcher;
+    }
   }
 }
 
 function replaceCredentialInResponse({
   definition,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions): void {
   if (definition.response) {
@@ -151,7 +131,6 @@ function replaceCredentialInResponse({
     response = replaceCredential({
       payload: response,
       credential,
-      isParameter,
       placeholder,
     });
 
@@ -162,25 +141,24 @@ function replaceCredentialInResponse({
 function replaceCredentialInScope({
   definition,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions): void {
-  debug('Replacing credentials in scope');
-  debugSensitive('Scope:', definition.scope);
+  if (definition.scope.includes(credential)) {
+    debug('Replacing credentials in scope');
+    debugSensitive('Scope:', definition.scope);
 
-  definition.scope = replaceCredential({
-    payload: definition.scope,
-    credential,
-    isParameter,
-    placeholder,
-  });
+    definition.scope = replaceCredential({
+      payload: definition.scope,
+      credential,
+      placeholder,
+    });
+  }
 }
 
 function replaceCredentialInQuery({
   definition,
   baseUrl,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions & { baseUrl: string }): void {
   const baseUrlOrigin = new URL(baseUrl).origin;
@@ -197,7 +175,6 @@ function replaceCredentialInQuery({
         replaceCredential({
           payload: queryValue,
           credential,
-          isParameter,
           placeholder,
         })
       );
@@ -212,7 +189,6 @@ function replaceCredentialInPath({
   definition,
   baseUrl,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions & { baseUrl: string }): void {
   const baseUrlOrigin = new URL(baseUrl).origin;
@@ -225,7 +201,6 @@ function replaceCredentialInPath({
     definitionURL.pathname = replaceCredential({
       payload: definitionURL.pathname,
       credential,
-      isParameter,
       placeholder,
     });
   }
@@ -238,7 +213,6 @@ function replaceApiKeyInHeader({
   definition,
   scheme,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions & { scheme: ApiKeySecurityScheme }): void {
   if (scheme.name !== undefined) {
@@ -253,7 +227,6 @@ function replaceApiKeyInHeader({
       definition.reqheaders[scheme.name] = replaceCredential({
         payload: definition.reqheaders[scheme.name].toString(),
         credential,
-        isParameter,
         placeholder,
       });
     }
@@ -262,7 +235,6 @@ function replaceApiKeyInHeader({
   replaceCredentialInHeaders({
     definition,
     credential,
-    isParameter,
     placeholder,
   });
 }
@@ -270,13 +242,11 @@ function replaceApiKeyInHeader({
 function replaceApiKeyInBody({
   definition,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions): void {
   replaceCredentialInBody({
     definition,
     credential,
-    isParameter,
     placeholder,
   });
 }
@@ -285,14 +255,12 @@ function replaceApiKeyInPath({
   definition,
   baseUrl,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions & { baseUrl: string }): void {
   replaceCredentialInPath({
     definition,
     baseUrl,
     credential,
-    isParameter,
     placeholder,
   });
 }
@@ -302,7 +270,6 @@ function replaceApiKeyInQuery({
   scheme,
   baseUrl,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions & { baseUrl: string; scheme: ApiKeySecurityScheme }): void {
   const baseUrlOrigin = new URL(baseUrl).origin;
@@ -310,28 +277,25 @@ function replaceApiKeyInQuery({
 
   if (
     scheme.name !== undefined &&
-    definitionURL.searchParams.has(scheme.name)
+    definitionURL.searchParams.has(scheme.name) &&
+    definitionURL.searchParams.get(scheme.name)?.includes(credential)
   ) {
     debug('Replacing api-key in query');
     debugSensitive('Query name:', scheme.name);
     debugSensitive('Query value:', definitionURL.searchParams.get(scheme.name));
 
-    definitionURL.searchParams.set(
-      scheme.name,
-      placeholder ? placeholder : defaultPlaceholder(isParameter)
-    );
+    definitionURL.searchParams.set(scheme.name, placeholder);
+
+    definition.path =
+      definitionURL.pathname + definitionURL.search + definitionURL.hash;
   }
 
   replaceCredentialInQuery({
     definition,
     baseUrl,
     credential,
-    isParameter,
     placeholder,
   });
-
-  definition.path =
-    definitionURL.pathname + definitionURL.search + definitionURL.hash;
 }
 
 function replaceApiKey({
@@ -339,14 +303,12 @@ function replaceApiKey({
   scheme,
   baseUrl,
   credential,
-  isParameter,
   placeholder,
 }: ReplaceOptions & { baseUrl: string; scheme: ApiKeySecurityScheme }): void {
   debug('Replacing api-key');
   const options = {
     definition,
     credential,
-    isParameter,
     placeholder,
   };
 
@@ -364,27 +326,23 @@ function replaceApiKey({
 
 function replaceBasicAuth(
   definition: RecordingDefinition,
-  placeholder?: string
+  placeholder: string
 ): void {
   if (definition.reqheaders?.[AUTH_HEADER_NAME] !== undefined) {
     debug('Replacing Basic token');
 
-    definition.reqheaders[AUTH_HEADER_NAME] = `Basic ${
-      placeholder ?? HIDDEN_CREDENTIALS_PLACEHOLDER
-    }`;
+    definition.reqheaders[AUTH_HEADER_NAME] = `Basic ${placeholder}`;
   }
 }
 
 function replaceBearerAuth(
   definition: RecordingDefinition,
-  placeholder?: string
+  placeholder: string
 ): void {
   if (definition.reqheaders?.[AUTH_HEADER_NAME] !== undefined) {
     debug('Replacing Bearer token');
 
-    definition.reqheaders[AUTH_HEADER_NAME] = `Bearer ${
-      placeholder ?? HIDDEN_CREDENTIALS_PLACEHOLDER
-    }`;
+    definition.reqheaders[AUTH_HEADER_NAME] = `Bearer ${placeholder}`;
   }
 }
 
@@ -399,14 +357,12 @@ export function replaceCredentialInDefinition({
   scheme: SecurityScheme;
   baseUrl: string;
   credential: string;
-  placeholder?: string;
+  placeholder: string;
 }): void {
   debug('Replacing credentials based on security schemes');
-  const isParameter = false;
   const options = {
     definition,
     credential,
-    isParameter,
     placeholder,
   };
 
@@ -443,14 +399,12 @@ export function replaceParameterInDefinition({
   definition: RecordingDefinition;
   baseUrl: string;
   credential: string;
-  placeholder?: string;
+  placeholder: string;
 }): void {
   debug('Replacing integration parameters');
-  const isParameter = true;
   const options = {
     definition,
     credential,
-    isParameter,
     placeholder,
   };