+++ src/superface-test.test.ts
@@ -20,7 +20,6 @@ import {
 import { matchWildCard } from './common/format';
 import { exists, readFileQuiet } from './common/io';
 import { writeRecordings } from './common/output-stream';
-import { HIDDEN_CREDENTIALS_PLACEHOLDER } from './nock.utils';
 import {
   getMockedSfConfig,
   getProfileMock,
@@ -29,6 +28,7 @@ import {
   SuperfaceClientMock,
 } from './superface.mock';
 import { SuperfaceTest } from './superface-test';
+import { HIDDEN_CREDENTIALS_PLACEHOLDER } from './superface-test.utils';
 
 /* eslint-disable @typescript-eslint/unbound-method */
 
@@ -272,7 +272,7 @@ describe('SuperfaceTest', () => {
           [
             {
               scope: 'https://localhost',
-              path: `/?api_key=${HIDDEN_CREDENTIALS_PLACEHOLDER}`,
+              path: `/?api_key=${HIDDEN_CREDENTIALS_PLACEHOLDER}api-key`,
               status: 200,
               response: {
                 some: 'data',