+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Warning about sensitive information before they're written
 - Hiding of credentials and parameters located in rawHeaders or in response.
 
+### Changed
+- Format of placeholders for sensitive information in recordings
+
 ### Fixed
 - Check for profile provider in super.json now does not expect defined profile